pyproject.toml
[tool.poetry]
name = "salenotify"
version = "0.1.0"
description = ""
authors = ["Dimon K Kazakhstan"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
aerich = "^0.7.2"
celery = "^5.3.1"
aioredis = "^2.0.1"
tortoise-orm = "^0.19.3"
pydantic-settings = "^2.0.2"
asyncpg = "^0.28.0"
structlog = "^23.1.0"
redis = "^4.6.0"
httpx = "^0.24.1"
beautifulsoup4 = "^4.12.2"
pytest-asyncio = "^0.21.1"
pytest-mock = "^3.11.1"
aiogram = "^2.25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.aerich]
location = "./src/migrations/"
src_folder = "./src/"
tortoise_orm = "config.TORTOISE_ORM"

